Refactored Task module to root
