+++ src/nwave/__init__.py
@@ -1,5 +1,5 @@
 __version__ = '0.1.0'
 
-from .core import WaveCore
-from .scheduler import Task, TaskResult, TaskException
 from .batch import Batch
+from .core import WaveCore
+from .task import Task, TaskResult, TaskException
